requirements.txt
streamlit>=1.36.0
openpyxl>=3.1.5
Pillow>=10.0.0
